Add gitattribute to prevent conflict in .gitignore I have two branches, A and B. Branch A have a directory examples with some files that are tracked by git, and these files should not appear on branch B. In my workflow, I do merge changes made in A into B often, which is a problem every time that there is some changes on examples. For the moment I am doing this manually: erasing the files after the merge or solving conflicts when there was a change to a file that I had already erased.
Solution:
On branch A:

Create a file named '.gitattributes' in the same dir, with this line:
database.yml merge=ours. This tells git what strategy to use when mergin
the file. In this case it always keep your version, ie. the version on
the branch you are merging into.

Add the .gitattributes file and commit

On branch B: repeat steps 1-2